I have attached all of the latest files. As you will notice, I have created a brand new Node class. Everything in there should be perfect. The Packet class should also be without any issues.
The Network class on the other hand, I have updated.

I created a generatePackets() method which generates packets for all
each station.
I also added a getSortedPacketList() which sorts array list of Packets.
In simulate(), I changed the WHILE condition to be false when
(currentTime <= MAX_TIME_SLOT)

In that while loop we need to be...
checking for collisionsdealing with collisions (backoff() )generating
new Poisson generated packets that we put into the queuecontinue these
steps

Now, the logic can be tweaked but this was my thought process.

NOTE : My initial idea of sending only after packets are ALL successful
is not correct.

NOTE :NOTE : I removed Part C code because it was getting sooo messy and
I couldn't read it any more. 

 I'm really sorry that I'm leaving you guys with this project!